+++ app/src/main/java/com/driot/bookplayer/player/ExoStreamPlayerEngine.java
@@ -172,7 +172,7 @@ public final class ExoStreamPlayerEngine extends LoggerHelper implements PlayerE
     @Override public boolean isPlaying() { return player != null && player.isPlaying(); }
     @Override public boolean isReady()   { return prepared && !preparing; }
 
-    @Override public int getCurrentPosition() {
+    @Override public long getCurrentPosition() {
         if (!prepared) {
             myLogD("getCurrentPosition() while not prepared -> 0");
             return 0;
@@ -181,7 +181,7 @@ public final class ExoStreamPlayerEngine extends LoggerHelper implements PlayerE
         catch (Throwable e) { myLogE("getCurrentPosition() ex"); return 0; }
     }
 
-    @Override public int getDuration() {
+    @Override public long getDuration() {
         if (!prepared) {
             myLogD("getDuration() while not prepared -> 0");
             return 0;
@@ -195,7 +195,7 @@ public final class ExoStreamPlayerEngine extends LoggerHelper implements PlayerE
         catch (Throwable ignored) { return 0; }
     }
 
-    @Override public void seekTo(int positionMs) {
+    @Override public void seekTo(long positionMs) {
         if (!prepared) {
             myLogE("seekTo(" + positionMs + ") while not prepared");
             return;
